Fix for Unit Tests taking too long: - only a limited set of QRDA patients are used - a separate set of tests with comprehensive coverage is in mRNfinder_masive.rb - the intention is to run this only as per need basis - with this change the whole test suite runs in less than 4 seconds on the cloud9 env.
